requirements.txt
Flask
python-docx
beautifulsoup4
